+++ examples/hover_shadow_html.rs
@@ -1,5 +1,5 @@
 fn main() -> core::fmt::Result {
-    let x = (0..50).map(|x| (x as f64 / 50.0) * 10.0);
+    let x = (0..50).map(|x| (x as f32 / 50.0) * 10.0);
 
     let mut s = poloto::plot("Demo: Hovering and shadows", "x", "y");
